Split fonts.rs and font.rs into one file per type
Pure move, no logic changes.
