+++ crates/epaint/src/text/fonts.rs
@@ -1,17 +1,19 @@
 use core::sync::atomic::{AtomicUsize, Ordering};
-use std::{borrow::Cow, collections::BTreeMap, sync::Arc};
+use std::{collections::BTreeMap, sync::Arc};
+
+use emath::NumExt as _;
 
 use crate::{
-    Color32, ColorImage, TextureAtlas,
+    Color32, TextureAtlas,
     text::{
-        ByteIndex, Galley, LayoutJob, LayoutSection, Tag, TextOptions, VariationCoords,
+        FontDefinitions, FontFamily, FontId, Galley, GlyphRasterizer, GlyphSource,
+        GlyphSourcePreference, LayoutJob, TextOptions, VariationCoords,
         font::{Font, FontFace, RasterGlyphAllocation, RasterGlyphCacheKey},
+        font_data::blob_from_font_data,
+        galley_cache::GalleyCache,
+        glyph_rasterizer::default_glyph_source,
     },
 };
-use emath::{NumExt as _, OrderedFloat, Rangef};
-
-#[cfg(feature = "default_fonts")]
-use epaint_default_fonts::{EMOJI_ICON, HACK_REGULAR, NOTO_EMOJI_REGULAR, UBUNTU_LIGHT};
 
 // ----------------------------------------------------------------------------
 
@@ -21,760 +23,6 @@ use epaint_default_fonts::{EMOJI_ICON, HACK_REGULAR, NOTO_EMOJI_REGULAR, UBUNTU_
 pub const MAX_GLYPH_SIZE: usize = 1024;
 
 // ----------------------------------------------------------------------------
-
-/// How to select a sized font.
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct FontId {
-    /// Height in points.
-    pub size: f32,
-
-    /// What font family to use.
-    pub family: FontFamily,
-    // TODO(emilk): weight (bold), italics, …
-}
-
-/// Input to a [`GlyphRasterizer`].
-pub struct GlyphRasterizerRequest<'a> {
-    /// An unsupported grapheme cluster.
-    pub cluster: &'a str,
-
-    /// The requested font family.
-    pub family: &'a FontFamily,
-
-    /// Requested font size in physical pixels.
-    pub font_size_px: f32,
-
-    /// Horizontal offset from the pixel grid in physical pixels.
-    pub subpixel_offset_px: f32,
-}
-
-/// A glyph rasterized by a platform fallback.
-pub struct RasterizedGlyph {
-    /// Pixels in physical pixels. Color glyphs retain their original colors.
-    pub image: ColorImage,
-
-    /// Offset from the baseline to the image top-left, in physical pixels.
-    pub offset_px: emath::Vec2,
-
-    /// Horizontal advance, in physical pixels.
-    pub advance_px: f32,
-
-    /// Do not tint this glyph with the text color.
-    pub is_color: bool,
-}
-
-type RasterizeFn =
-    dyn for<'a> Fn(&GlyphRasterizerRequest<'a>) -> Option<RasterizedGlyph> + Send + Sync;
-
-/// Where to look first for the glyphs of a grapheme cluster.
-///
-/// The other source is used as a fallback if the first one cannot render the cluster.
-///
-/// See [`GlyphSourcePreference`].
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum GlyphSource {
-    /// The fonts in [`FontDefinitions`].
-    ///
-    /// Predictable: looks the same everywhere, both on native and on web.
-    Fonts,
-
-    /// What the platform offers, e.g. the [`GlyphRasterizer`] (the browser on web).
-    ///
-    /// Supports colored emojis.
-    /// Unpredictable: may look different on different computers.
-    Platform,
-}
-
-/// Decides where to look first for the glyphs of each grapheme cluster.
-///
-/// Default: [`default_glyph_source`], so that color emoji come from the platform,
-/// while text-presentation symbols (e.g. ⏮︎) look the same on all platforms.
-///
-/// Use `|_| GlyphSource::Fonts` to only use the platform for clusters
-/// that no font in [`FontDefinitions`] can render.
-///
-/// Set with `egui::Context::set_glyph_source_preference` or [`Fonts::with_glyph_source_preference`].
-pub type GlyphSourcePreference = Arc<dyn Fn(&str) -> GlyphSource + Send + Sync>;
-
-/// Rasterizes grapheme clusters using something other than the installed fonts,
-/// e.g. the browser on web.
-///
-/// Used for clusters no installed font can render,
-/// and for clusters where the [`GlyphSourcePreference`] says [`GlyphSource::Platform`].
-#[derive(Clone)]
-pub struct GlyphRasterizer {
-    /// Rasterize one grapheme cluster.
-    ///
-    /// Return `None` if the platform cannot render it either.
-    pub rasterize: Arc<RasterizeFn>,
-}
-
-impl GlyphRasterizer {
-    pub fn new(
-        rasterize: impl for<'a> Fn(&GlyphRasterizerRequest<'a>) -> Option<RasterizedGlyph>
-        + Send
-        + Sync
-        + 'static,
-    ) -> Self {
-        Self {
-            rasterize: Arc::new(rasterize),
-        }
-    }
-}
-
-/// The default [`GlyphSourcePreference`]:
-/// [`GlyphSource::Platform`] for clusters with emoji presentation
-/// (see [`has_emoji_presentation`]), [`GlyphSource::Fonts`] for everything else.
-pub fn default_glyph_source(cluster: &str) -> GlyphSource {
-    if has_emoji_presentation(cluster) {
-        GlyphSource::Platform
-    } else {
-        GlyphSource::Fonts
-    }
-}
-
-impl core::fmt::Debug for GlyphRasterizer {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.write_str("GlyphRasterizer")
-    }
-}
-
-/// Does this grapheme cluster have emoji presentation, per Unicode (UTS #51)?
-///
-/// True for clusters that default to a color glyph (😀, 🦀, 🇸🇪, 👨‍👩‍👧),
-/// for clusters with an explicit emoji presentation selector (⏮️, U+FE0F),
-/// and for emoji modifier (skin tone) sequences (☝🏻).
-///
-/// False for text-presentation symbols (⏮, ✔, ♥) and for clusters
-/// with an explicit text presentation selector (⏮︎, U+FE0E).
-///
-/// Used by [`default_glyph_source`].
-pub fn has_emoji_presentation(cluster: &str) -> bool {
-    use unicode_properties::emoji::{
-        EmojiStatus, UnicodeEmoji as _, is_emoji_presentation_selector,
-        is_text_presentation_selector,
-    };
-
-    if cluster.is_ascii() {
-        return false; // Fast path: no ASCII character has emoji presentation.
-    }
-
-    let mut has_emoji_presentation = false;
-    for c in cluster.chars() {
-        if is_text_presentation_selector(c) {
-            return false; // Explicit text presentation wins.
-        }
-        has_emoji_presentation |= is_emoji_presentation_selector(c)
-            || matches!(
-                c.emoji_status(),
-                EmojiStatus::EmojiPresentation
-                    | EmojiStatus::EmojiPresentationAndModifierBase
-                    | EmojiStatus::EmojiPresentationAndEmojiComponent
-                    | EmojiStatus::EmojiPresentationAndModifierAndEmojiComponent
-            );
-    }
-    has_emoji_presentation
-}
-
-#[cfg(test)]
-mod has_emoji_presentation_tests {
-    use super::has_emoji_presentation;
-
-    #[test]
-    fn emoji_presentation() {
-        for cluster in [
-            "😀",
-            "🦀",
-            "⏮\u{FE0F}",              // explicit emoji presentation
-            "☝🏻",                     // skin tone modifier sequence
-            "🇸🇪",                     // flag
-            "👨\u{200D}👩\u{200D}👧", // ZWJ sequence
-            "1\u{FE0F}\u{20E3}",      // keycap
-        ] {
-            assert!(has_emoji_presentation(cluster), "{cluster:?}");
-        }
-    }
-
-    #[test]
-    fn text_presentation() {
-        for cluster in [
-            "",
-            "a",
-            "1",
-            "⏮",
-            "⏮\u{FE0E}",  // explicit text presentation
-            "😀\u{FE0E}", // explicit text presentation wins
-            "✔",
-            "♥",
-            "©",
-            "→",
-            "√",
-        ] {
-            assert!(!has_emoji_presentation(cluster), "{cluster:?}");
-        }
-    }
-}
-
-impl Default for FontId {
-    #[inline]
-    fn default() -> Self {
-        Self {
-            size: 14.0,
-            family: FontFamily::Proportional,
-        }
-    }
-}
-
-impl FontId {
-    #[inline]
-    pub const fn new(size: f32, family: FontFamily) -> Self {
-        Self { size, family }
-    }
-
-    #[inline]
-    pub const fn proportional(size: f32) -> Self {
-        Self::new(size, FontFamily::Proportional)
-    }
-
-    #[inline]
-    pub const fn monospace(size: f32) -> Self {
-        Self::new(size, FontFamily::Monospace)
-    }
-}
-
-impl core::hash::Hash for FontId {
-    #[inline(always)]
-    fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
-        let Self { size, family } = self;
-        emath::OrderedFloat(*size).hash(state);
-        family.hash(state);
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// Font of unknown size.
-///
-/// Which style of font: [`Monospace`][`FontFamily::Monospace`], [`Proportional`][`FontFamily::Proportional`],
-/// or by user-chosen name.
-#[derive(Clone, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub enum FontFamily {
-    /// A font where some characters are wider than other (e.g. 'w' is wider than 'i').
-    ///
-    /// Proportional fonts are easier to read and should be the preferred choice in most situations.
-    #[default]
-    Proportional,
-
-    /// A font where each character is the same width (`w` is the same width as `i`).
-    ///
-    /// Useful for code snippets, or when you need to align numbers or text.
-    Monospace,
-
-    /// One of the names in [`FontDefinitions::families`].
-    ///
-    /// ```
-    /// # use epaint::FontFamily;
-    /// // User-chosen names:
-    /// FontFamily::Name("arial".into());
-    /// FontFamily::Name("serif".into());
-    /// ```
-    Name(Arc<str>),
-}
-
-impl core::fmt::Display for FontFamily {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        match self {
-            Self::Monospace => "Monospace".fmt(f),
-            Self::Proportional => "Proportional".fmt(f),
-            Self::Name(name) => (*name).fmt(f),
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// A `.ttf` or `.otf` file and a font face index.
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct FontData {
-    /// The content of a `.ttf` or `.otf` file.
-    pub font: Cow<'static, [u8]>,
-
-    /// Which font face in the file to use.
-    /// When in doubt, use `0`.
-    pub index: u32,
-
-    /// Extra scale and vertical tweak to apply to all text of this font.
-    pub tweak: FontTweak,
-}
-
-impl FontData {
-    pub fn from_static(font: &'static [u8]) -> Self {
-        Self {
-            font: Cow::Borrowed(font),
-            index: 0,
-            tweak: Default::default(),
-        }
-    }
-
-    pub fn from_owned(font: Vec<u8>) -> Self {
-        Self {
-            font: Cow::Owned(font),
-            index: 0,
-            tweak: Default::default(),
-        }
-    }
-
-    pub fn tweak(self, tweak: FontTweak) -> Self {
-        Self { tweak, ..self }
-    }
-
-    /// The variation axes of this font, e.g. `wght` (weight) and `wdth` (width).
-    ///
-    /// Use this to discover which axes a variable font supports, and their valid
-    /// ranges, so a UI can offer the right knobs instead of making the user guess
-    /// tags and values for [`FontTweak::coords`].
-    ///
-    /// Returns an empty list for non-variable (static) fonts, or if the font data
-    /// fails to parse.
-    pub fn variation_axes(&self) -> Vec<FontVariationAxis> {
-        use skrifa::MetadataProvider as _;
-
-        let Ok(font) = skrifa::FontRef::from_index(self.font.as_ref(), self.index) else {
-            return Vec::new();
-        };
-
-        font.axes()
-            .iter()
-            .map(|axis| FontVariationAxis {
-                tag: axis.tag(),
-                name: font
-                    .localized_strings(axis.name_id())
-                    .english_or_first()
-                    .map(|name| name.chars().collect()),
-                range: Rangef::new(axis.min_value(), axis.max_value()),
-                default: axis.default_value(),
-                hidden: axis.is_hidden(),
-            })
-            .collect()
-    }
-}
-
-/// A single variation axis of a variable font, e.g. weight (`wght`) or width (`wdth`).
-///
-/// Obtained via [`FontData::variation_axes`].
-#[derive(Clone, Debug, PartialEq)]
-pub struct FontVariationAxis {
-    /// The axis tag, e.g. `wght` or `wdth`.
-    pub tag: Tag,
-
-    /// Human-readable axis name, if the font provides one (e.g. "Weight").
-    pub name: Option<String>,
-
-    /// Valid range of values for this axis, `min..=max`.
-    pub range: Rangef,
-
-    /// The value used when the axis is not overridden.
-    pub default: f32,
-
-    /// Whether the font recommends hiding this axis from user interfaces.
-    pub hidden: bool,
-}
-
-impl AsRef<[u8]> for FontData {
-    fn as_ref(&self) -> &[u8] {
-        self.font.as_ref()
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// Extra scale and vertical tweak to apply to all text of a certain font.
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct FontTweak {
-    /// Scale the font's glyphs by this much.
-    /// this is only a visual effect and does not affect the text layout.
-    ///
-    /// Default: `1.0` (no scaling).
-    pub scale: f32,
-
-    /// Shift font's glyphs downwards by this fraction of the font size (in points).
-    /// this is only a visual effect and does not affect the text layout.
-    ///
-    /// Affects larger font sizes more.
-    ///
-    /// A positive value shifts the text downwards.
-    /// A negative value shifts it upwards.
-    ///
-    /// Example value: `-0.2`.
-    pub y_offset_factor: f32,
-
-    /// Shift font's glyphs downwards by this amount of logical points.
-    /// this is only a visual effect and does not affect the text layout.
-    ///
-    /// Affects all font sizes equally.
-    ///
-    /// Example value: `2.0`.
-    pub y_offset: f32,
-
-    /// Override the global font hinting setting for this specific font.
-    ///
-    /// `None` means use the global setting in [`TextOptions::font_hinting`].
-    pub hinting: Option<bool>,
-
-    /// How to grid-fit the glyph outlines when hinting is enabled.
-    ///
-    /// Has no effect when hinting is disabled (see [`Self::hinting`]).
-    pub hinting_target: HintingTarget,
-
-    /// Override the global sub-pixel binning setting for this specific font.
-    ///
-    /// `None` means use the global setting in [`TextOptions::subpixel_binning`].
-    pub subpixel_binning: Option<bool>,
-
-    /// Override the font's default variation coordinates for its axes ("wght", etc.).
-    pub coords: VariationCoords,
-
-    /// Width of a thin space (`\u{2009}`) and narrow no-break space (`\u{202F}`),
-    /// as a fraction of the normal space width.
-    ///
-    /// Thin space is often used as a thousands separator: `1 234 567`.
-    ///
-    /// Default: `0.5` (half a normal space).
-    pub thin_space_width: f32,
-
-    /// Width of a tab character (`\t`), measured in number of space widths.
-    ///
-    /// Default: `4.0`.
-    pub tab_size: f32,
-}
-
-impl Default for FontTweak {
-    fn default() -> Self {
-        Self {
-            scale: 1.0,
-            y_offset_factor: 0.0,
-            y_offset: 0.0,
-            hinting: None,
-            hinting_target: HintingTarget::default(),
-            subpixel_binning: None,
-            coords: VariationCoords::default(),
-            thin_space_width: 0.5,
-            tab_size: 4.0,
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// How to *hint* glyph outlines, i.e. how aggressively to nudge them onto the
-/// pixel grid before rasterizing. Mirrors [`skrifa::outline::Target`].
-///
-/// Hinting trades shape fidelity for sharpness: snapping stems to whole pixels
-/// makes text crisp at small sizes / low dpi, at the cost of slightly distorting
-/// the designer's outlines. At high dpi it matters little.
-///
-/// This only has an effect if the font is actually hinted — either it ships
-/// TrueType instructions, or it was auto-hinted (see [`FontTweak::hinting`]).
-///
-/// Used by [`FontTweak::hinting_target`].
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub enum HintingTarget {
-    /// Strongest hinting, designed for aliased 1-bit (black & white) rendering.
-    ///
-    /// Snaps stems hard to the pixel grid for maximum sharpness. egui always
-    /// renders anti-aliased, so in practice this looks much like [`Self::Smooth`]
-    /// here; it mostly exists for completeness. Maps to `skrifa`'s `Target::Mono`.
-    Mono,
-
-    /// Hinting tuned for anti-aliased rendering. This is what you normally want,
-    /// and what egui uses by default. Maps to `skrifa`'s `Target::Smooth`.
-    Smooth(SmoothHinting),
-}
-
-impl Default for HintingTarget {
-    fn default() -> Self {
-        Self::Smooth(SmoothHinting::default())
-    }
-}
-
-impl From<HintingTarget> for skrifa::outline::Target {
-    fn from(hinting_target: HintingTarget) -> Self {
-        use skrifa::outline::SmoothMode;
-        match hinting_target {
-            HintingTarget::Mono => Self::Mono,
-            HintingTarget::Smooth(SmoothHinting {
-                light,
-                symmetric_rendering,
-                preserve_linear_metrics,
-            }) => Self::Smooth {
-                mode: if light {
-                    SmoothMode::Light
-                } else {
-                    SmoothMode::Normal
-                },
-                symmetric_rendering,
-                preserve_linear_metrics,
-            },
-        }
-    }
-}
-
-/// Tuning for [`HintingTarget::Smooth`], mirroring `skrifa`'s `Target::Smooth`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct SmoothHinting {
-    /// Hint only lightly: snap stems vertically but leave horizontal shapes
-    /// alone (`FreeType`'s "light" mode). Preserves the font's proportions
-    /// better, at the cost of a little horizontal sharpness.
-    ///
-    /// `false` uses the "normal" mode, which also fits horizontally.
-    /// Maps to `SmoothMode::Light` (`true`) vs `SmoothMode::Normal` (`false`).
-    pub light: bool,
-
-    /// Render a glyph the same way regardless of its sub-pixel position.
-    ///
-    /// `true` makes glyphs position-independent (better for caching and
-    /// animation), but a font's instructions may then widen stems and look
-    /// slightly blurrier under an analytic rasterizer like egui's.
-    ///
-    /// **Only affects fonts hinted via the TrueType interpreter** (i.e. fonts
-    /// that ship their own instructions). It has no effect on the auto-hinter.
-    /// Mirrors `Target::Smooth { symmetric_rendering }`.
-    pub symmetric_rendering: bool,
-
-    /// Keep advance widths independent of hinting (don't grid-fit horizontally
-    /// in a way that changes spacing).
-    ///
-    /// `true` keeps inter-glyph spacing identical to the unhinted font, so
-    /// layout never depends on hinting — but it also prevents horizontal
-    /// grid-fitting, leaving vertical stems softer on low-dpi screens.
-    ///
-    /// `false` lets the (auto)hinter snap horizontally for crisper stems.
-    /// egui positions glyphs from the shaper's advances, not the hinted
-    /// outline, so this mainly affects sharpness here, not layout.
-    /// Mirrors `Target::Smooth { preserve_linear_metrics }`.
-    pub preserve_linear_metrics: bool,
-}
-
-impl Default for SmoothHinting {
-    fn default() -> Self {
-        // Matches the behavior egui had before the hinting target was configurable.
-        // Note this means horizontal grid-fitting is opt-in (see `preserve_linear_metrics`).
-        Self {
-            light: false,
-            symmetric_rendering: true,
-            preserve_linear_metrics: true,
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-pub type Blob = Arc<dyn AsRef<[u8]> + Send + Sync>;
-
-fn blob_from_font_data(data: &FontData) -> Blob {
-    match data.clone().font {
-        Cow::Borrowed(bytes) => Arc::new(bytes) as Blob,
-        Cow::Owned(bytes) => Arc::new(bytes) as Blob,
-    }
-}
-
-/// Describes the font data and the sizes to use.
-///
-/// Often you would start with [`FontDefinitions::default()`] and then add/change the contents.
-///
-/// This is how you install your own custom fonts:
-/// ```
-/// # use {epaint::text::{FontDefinitions, FontFamily, FontData}};
-/// # struct FakeEguiCtx {};
-/// # impl FakeEguiCtx { fn set_fonts(&self, _: FontDefinitions) {} }
-/// # let egui_ctx = FakeEguiCtx {};
-/// let mut fonts = FontDefinitions::default();
-///
-/// // Install my own font (maybe supporting non-latin characters):
-/// fonts.font_data.insert("my_font".to_owned(),
-///    std::sync::Arc::new(
-///        // .ttf and .otf supported
-///        FontData::from_static(include_bytes!("../../../epaint_default_fonts/fonts/Ubuntu-Light.ttf"))
-///    )
-/// );
-///
-/// // Put my font first (highest priority):
-/// fonts.families.get_mut(&FontFamily::Proportional).unwrap()
-///     .insert(0, "my_font".to_owned());
-///
-/// // Put my font as last fallback for monospace:
-/// fonts.families.get_mut(&FontFamily::Monospace).unwrap()
-///     .push("my_font".to_owned());
-///
-/// egui_ctx.set_fonts(fonts);
-/// ```
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "serde", serde(default))]
-pub struct FontDefinitions {
-    /// List of font names and their definitions.
-    ///
-    /// `epaint` has built-in-default for these, but you can override them if you like.
-    pub font_data: BTreeMap<String, Arc<FontData>>,
-
-    /// Which fonts (names) to use for each [`FontFamily`].
-    ///
-    /// The list should be a list of keys into [`Self::font_data`].
-    /// When looking for a character glyph `epaint` will start with
-    /// the first font and then move to the second, and so on.
-    /// So the first font is the primary, and then comes a list of fallbacks in order of priority.
-    pub families: BTreeMap<FontFamily, Vec<String>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct FontInsert {
-    /// Font name
-    pub name: String,
-
-    /// A `.ttf` or `.otf` file and a font face index.
-    pub data: FontData,
-
-    /// Sets the font family and priority
-    pub families: Vec<InsertFontFamily>,
-}
-
-#[derive(Debug, Clone)]
-pub struct InsertFontFamily {
-    /// Font family
-    pub family: FontFamily,
-
-    /// Fallback or Primary font
-    pub priority: FontPriority,
-}
-
-#[derive(Debug, Clone)]
-pub enum FontPriority {
-    /// Prefer this font before all existing ones.
-    ///
-    /// If a desired glyph exists in this font, it will be used.
-    Highest,
-
-    /// Use this font as a fallback, after all existing ones.
-    ///
-    /// This font will only be used if the glyph is not found in any of the previously installed fonts.
-    Lowest,
-}
-
-impl FontInsert {
-    pub fn new(name: &str, data: FontData, families: Vec<InsertFontFamily>) -> Self {
-        Self {
-            name: name.to_owned(),
-            data,
-            families,
-        }
-    }
-}
-
-impl Default for FontDefinitions {
-    /// Specifies the default fonts if the feature `default_fonts` is enabled,
-    /// otherwise this is the same as [`Self::empty`].
-    #[cfg(not(feature = "default_fonts"))]
-    fn default() -> Self {
-        Self::empty()
-    }
-
-    /// Specifies the default fonts if the feature `default_fonts` is enabled,
-    /// otherwise this is the same as [`Self::empty`].
-    #[cfg(feature = "default_fonts")]
-    fn default() -> Self {
-        let mut font_data: BTreeMap<String, Arc<FontData>> = BTreeMap::new();
-
-        let mut families = BTreeMap::new();
-
-        font_data.insert(
-            "Hack".to_owned(),
-            Arc::new(FontData::from_static(HACK_REGULAR)),
-        );
-
-        // Some good looking emojis. Use as first priority:
-        font_data.insert(
-            "NotoEmoji-Regular".to_owned(),
-            Arc::new(FontData::from_static(NOTO_EMOJI_REGULAR).tweak(FontTweak {
-                scale: 0.81, // Make smaller
-                ..Default::default()
-            })),
-        );
-
-        font_data.insert(
-            "Ubuntu-Light".to_owned(),
-            Arc::new(FontData::from_static(UBUNTU_LIGHT)),
-        );
-
-        // Bigger emojis, and more. <http://jslegers.github.io/emoji-icon-font/>:
-        font_data.insert(
-            "emoji-icon-font".to_owned(),
-            Arc::new(FontData::from_static(EMOJI_ICON).tweak(FontTweak {
-                scale: 0.90, // Make smaller
-                ..Default::default()
-            })),
-        );
-
-        families.insert(
-            FontFamily::Monospace,
-            vec![
-                "Hack".to_owned(),
-                "Ubuntu-Light".to_owned(), // fallback for √ etc
-                "NotoEmoji-Regular".to_owned(),
-                "emoji-icon-font".to_owned(),
-            ],
-        );
-        families.insert(
-            FontFamily::Proportional,
-            vec![
-                "Ubuntu-Light".to_owned(),
-                "NotoEmoji-Regular".to_owned(),
-                "emoji-icon-font".to_owned(),
-            ],
-        );
-
-        Self {
-            font_data,
-            families,
-        }
-    }
-}
-
-impl FontDefinitions {
-    /// No fonts.
-    pub fn empty() -> Self {
-        let mut families = BTreeMap::new();
-        families.insert(FontFamily::Monospace, vec![]);
-        families.insert(FontFamily::Proportional, vec![]);
-
-        Self {
-            font_data: Default::default(),
-            families,
-        }
-    }
-
-    /// List of all the builtin font names used by `epaint`.
-    #[cfg(feature = "default_fonts")]
-    pub fn builtin_font_names() -> &'static [&'static str] {
-        &[
-            "Ubuntu-Light",
-            "NotoEmoji-Regular",
-            "emoji-icon-font",
-            "Hack",
-        ]
-    }
-
-    /// List of all the builtin font names used by `epaint`.
-    #[cfg(not(feature = "default_fonts"))]
-    pub fn builtin_font_names() -> &'static [&'static str] {
-        &[]
-    }
-}
 
 /// Unique ID for looking up a single font face/file.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
@@ -881,6 +129,8 @@ impl CachedFamily {
         None
     }
 }
+
+// ----------------------------------------------------------------------------
 
 // ----------------------------------------------------------------------------
 
@@ -1318,474 +568,10 @@ impl FontsImpl {
     }
 }
 
-// ----------------------------------------------------------------------------
-
-struct CachedGalley {
-    /// When it was last used
-    last_used: u32,
-
-    /// Hashes of all other entries this one depends on for quick re-layout.
-    /// Their `last_used`s should be updated alongside this one to make sure they're
-    /// not evicted.
-    children: Option<Arc<[u64]>>,
-
-    galley: Arc<Galley>,
-}
-
-#[derive(Default)]
-struct GalleyCache {
-    /// Frame counter used to do garbage collection on the cache
-    generation: u32,
-    cache: nohash_hasher::IntMap<u64, CachedGalley>,
-}
-
-impl GalleyCache {
-    fn layout_internal(
-        &mut self,
-        fonts: &mut FontsImpl,
-        mut job: LayoutJob,
-        pixels_per_point: f32,
-        allow_split_paragraphs: bool,
-    ) -> (u64, Arc<Galley>) {
-        if job.wrap.max_width.is_finite() {
-            // Protect against rounding errors in egui layout code.
-
-            // Say the user asks to wrap at width 200.0.
-            // The text layout wraps, and reports that the final width was 196.0 points.
-            // This then trickles up the `Ui` chain and gets stored as the width for a tooltip (say).
-            // On the next frame, this is then set as the max width for the tooltip,
-            // and we end up calling the text layout code again, this time with a wrap width of 196.0.
-            // Except, somewhere in the `Ui` chain with added margins etc, a rounding error was introduced,
-            // so that we actually set a wrap-width of 195.9997 instead.
-            // Now the text that fit perfrectly at 196.0 needs to wrap one word earlier,
-            // and so the text re-wraps and reports a new width of 185.0 points.
-            // And then the cycle continues.
-
-            // So we limit max_width to integers.
-
-            // Related issues:
-            // * https://github.com/emilk/egui/issues/4927
-            // * https://github.com/emilk/egui/issues/4928
-            // * https://github.com/emilk/egui/issues/5084
-            // * https://github.com/emilk/egui/issues/5163
-
-            job.wrap.max_width = job.wrap.max_width.round();
-        }
-
-        let hash = crate::util::hash((&job, OrderedFloat(pixels_per_point))); // TODO(emilk): even faster hasher?
-
-        let galley = match self.cache.entry(hash) {
-            std::collections::hash_map::Entry::Occupied(entry) => {
-                // The job was found in cache - no need to re-layout.
-                let cached = entry.into_mut();
-                cached.last_used = self.generation;
-
-                let galley = Arc::clone(&cached.galley);
-                if let Some(children) = &cached.children {
-                    // The point of `allow_split_paragraphs` is to split large jobs into paragraph,
-                    // and then cache each paragraph individually.
-                    // That way, if we edit a single paragraph, only that paragraph will be re-layouted.
-                    // For that to work we need to keep all the child/paragraph
-                    // galleys alive while the parent galley is alive:
-                    for child_hash in Arc::clone(children).iter() {
-                        if let Some(cached_child) = self.cache.get_mut(child_hash) {
-                            cached_child.last_used = self.generation;
-                        }
-                    }
-                }
-
-                galley
-            }
-            std::collections::hash_map::Entry::Vacant(entry) => {
-                let job = Arc::new(job);
-                if allow_split_paragraphs && should_cache_each_paragraph_individually(&job) {
-                    let (child_galleys, child_hashes) =
-                        self.layout_each_paragraph_individually(fonts, &job, pixels_per_point);
-                    debug_assert_eq!(
-                        child_hashes.len(),
-                        child_galleys.len(),
-                        "Bug in `layout_each_paragraph_individually`"
-                    );
-                    let galley = Arc::new(Galley::concat(job, &child_galleys, pixels_per_point));
-
-                    self.cache.insert(
-                        hash,
-                        CachedGalley {
-                            last_used: self.generation,
-                            children: Some(child_hashes.into()),
-                            galley: Arc::clone(&galley),
-                        },
-                    );
-                    galley
-                } else {
-                    let galley = super::layout(fonts, pixels_per_point, job);
-                    let galley = Arc::new(galley);
-                    entry.insert(CachedGalley {
-                        last_used: self.generation,
-                        children: None,
-                        galley: Arc::clone(&galley),
-                    });
-                    galley
-                }
-            }
-        };
-
-        (hash, galley)
-    }
-
-    fn layout(
-        &mut self,
-        fonts: &mut FontsImpl,
-        pixels_per_point: f32,
-        job: LayoutJob,
-        allow_split_paragraphs: bool,
-    ) -> Arc<Galley> {
-        self.layout_internal(fonts, job, pixels_per_point, allow_split_paragraphs)
-            .1
-    }
-
-    /// Split on `\n` and lay out (and cache) each paragraph individually.
-    fn layout_each_paragraph_individually(
-        &mut self,
-        fonts: &mut FontsImpl,
-        job: &LayoutJob,
-        pixels_per_point: f32,
-    ) -> (Vec<Arc<Galley>>, Vec<u64>) {
-        profiling::function_scope!();
-
-        let mut current_section = 0;
-        let mut start = 0;
-        let mut max_rows_remaining = job.wrap.max_rows;
-        let mut child_galleys = Vec::new();
-        let mut child_hashes = Vec::new();
-
-        while start < job.text.len() {
-            let is_first_paragraph = start == 0;
-            // `end` will not include the `\n` since we don't want to create an empty row in our
-            // split galley
-            let mut end = job.text[start..]
-                .find('\n')
-                .map_or(job.text.len(), |i| start + i);
-            if end == job.text.len() - 1 && job.text.ends_with('\n') {
-                end += 1; // If the text ends with a newline, we include it in the last paragraph.
-            }
-
-            let mut paragraph_job = LayoutJob {
-                text: job.text[start..end].to_owned(),
-                wrap: crate::text::TextWrapping {
-                    max_rows: max_rows_remaining,
-                    ..job.wrap
-                },
-                sections: Vec::new(),
-                break_on_newline: job.break_on_newline,
-                halign: job.halign,
-                justify: job.justify,
-                first_row_min_height: if is_first_paragraph {
-                    job.first_row_min_height
-                } else {
-                    0.0
-                },
-                round_output_to_gui: job.round_output_to_gui,
-                keep_trailing_whitespace: job.keep_trailing_whitespace,
-            };
-
-            // Add overlapping sections:
-            for section in &job.sections[current_section..job.sections.len()] {
-                let LayoutSection {
-                    leading_space,
-                    byte_range: section_range,
-                    format,
-                } = section;
-
-                // `start` and `end` are the byte range of the current paragraph.
-                // How does the current section overlap with the paragraph range?
-
-                if section_range.end <= ByteIndex(start) {
-                    // The section is behind us
-                    current_section += 1;
-                } else if ByteIndex(end) < section_range.start {
-                    break; // Haven't reached this one yet.
-                } else {
-                    // Section range overlaps with paragraph range
-                    debug_assert!(
-                        section_range.start <= section_range.end,
-                        "Bad byte_range: {section_range:?}"
-                    );
-                    let new_range = section_range.start.saturating_sub(start)
-                        ..(section_range.end.min(ByteIndex(end))).saturating_sub(start);
-                    debug_assert!(
-                        new_range.start <= new_range.end,
-                        "Bad new section range: {new_range:?}"
-                    );
-                    paragraph_job.sections.push(LayoutSection {
-                        leading_space: if ByteIndex(start) <= section_range.start {
-                            *leading_space
-                        } else {
-                            0.0
-                        },
-                        byte_range: new_range,
-                        format: format.clone(),
-                    });
-                }
-            }
-
-            // TODO(emilk): we could lay out each paragraph in parallel to get a nice speedup on multicore machines.
-            let (hash, galley) =
-                self.layout_internal(fonts, paragraph_job, pixels_per_point, false);
-            child_hashes.push(hash);
-
-            // This will prevent us from invalidating cache entries unnecessarily:
-            if max_rows_remaining != usize::MAX {
-                max_rows_remaining -= galley.rows.len();
-            }
-
-            let elided = galley.elided;
-            child_galleys.push(galley);
-            if elided {
-                break;
-            }
-
-            start = end + 1;
-        }
-
-        (child_galleys, child_hashes)
-    }
-
-    pub fn num_galleys_in_cache(&self) -> usize {
-        self.cache.len()
-    }
-
-    /// Must be called once per frame to clear the [`Galley`] cache.
-    pub fn flush_cache(&mut self) {
-        let current_generation = self.generation;
-        self.cache.retain(|_key, cached| {
-            cached.last_used == current_generation // only keep those that were used this frame
-        });
-        self.generation = self.generation.wrapping_add(1);
-    }
-}
-
-/// If true, lay out and cache each paragraph (sections separated by newlines) individually.
-///
-/// This makes it much faster to re-layout the full text when only a portion of it has changed since last frame, i.e. when editing somewhere in a file with thousands of lines/paragraphs.
-fn should_cache_each_paragraph_individually(job: &LayoutJob) -> bool {
-    // We currently don't support this elided text, i.e. when `max_rows` is set.
-    // Most often, elided text is elided to one row,
-    // and so will always be fast to lay out.
-    job.break_on_newline && job.wrap.max_rows == usize::MAX && job.text.contains('\n')
-}
-
 #[cfg(feature = "default_fonts")]
 #[cfg(test)]
 mod tests {
-    use core::f32;
-
     use super::*;
-    use crate::text::{TextWrapping, layout};
-    use crate::{Stroke, text::TextFormat};
-    use ecolor::Color32;
-    use emath::Align;
-
-    fn jobs() -> Vec<LayoutJob> {
-        vec![
-            LayoutJob::simple(
-                String::default(),
-                FontId::new(14.0, FontFamily::Monospace),
-                Color32::WHITE,
-                f32::INFINITY,
-            ),
-            LayoutJob::simple(
-                "ends with newlines\n\n".to_owned(),
-                FontId::new(14.0, FontFamily::Monospace),
-                Color32::WHITE,
-                f32::INFINITY,
-            ),
-            LayoutJob::simple(
-                "Simple test.".to_owned(),
-                FontId::new(14.0, FontFamily::Monospace),
-                Color32::WHITE,
-                f32::INFINITY,
-            ),
-            {
-                let mut job = LayoutJob::simple(
-                    "hi".to_owned(),
-                    FontId::default(),
-                    Color32::WHITE,
-                    f32::INFINITY,
-                );
-                job.append("\n", 0.0, TextFormat::default());
-                job.append("\n", 0.0, TextFormat::default());
-                job.append("world", 0.0, TextFormat::default());
-                job.wrap.max_rows = 2;
-                job
-            },
-            {
-                let mut job = LayoutJob::simple(
-                    "Test text with a lot of words\n and a newline.".to_owned(),
-                    FontId::new(14.0, FontFamily::Monospace),
-                    Color32::WHITE,
-                    40.0,
-                );
-                job.first_row_min_height = 30.0;
-                job
-            },
-            LayoutJob::simple(
-                "This some text that may be long.\nDet kanske också finns lite ÅÄÖ här.".to_owned(),
-                FontId::new(14.0, FontFamily::Proportional),
-                Color32::WHITE,
-                50.0,
-            ),
-            {
-                let mut job = LayoutJob {
-                    first_row_min_height: 20.0,
-                    ..Default::default()
-                };
-                job.append(
-                    "1st paragraph has underline and strikethrough, and has some non-ASCII characters:\n ÅÄÖ.",
-                    0.0,
-                    TextFormat {
-                        font_id: FontId::new(15.0, FontFamily::Monospace),
-                        underline: Stroke::new(1.0, Color32::RED),
-                        strikethrough: Stroke::new(1.0, Color32::GREEN),
-                        ..Default::default()
-                    },
-                );
-                job.append(
-                    "2nd paragraph has some leading space.\n",
-                    16.0,
-                    TextFormat {
-                        font_id: FontId::new(14.0, FontFamily::Proportional),
-                        ..Default::default()
-                    },
-                );
-                job.append(
-                    "3rd paragraph is kind of boring, but has italics.\nAnd a newline",
-                    0.0,
-                    TextFormat {
-                        font_id: FontId::new(10.0, FontFamily::Proportional),
-                        italics: true,
-                        ..Default::default()
-                    },
-                );
-
-                job
-            },
-            {
-                // Regression test for <https://github.com/emilk/egui/issues/7378>
-                let mut job = LayoutJob::default();
-                job.append("\n", 0.0, TextFormat::default());
-                job.append("", 0.0, TextFormat::default());
-                job
-            },
-        ]
-    }
-
-    #[expect(clippy::print_stdout)]
-    #[test]
-    fn test_split_paragraphs() {
-        for pixels_per_point in [1.0, 2.0_f32.sqrt(), 2.0] {
-            let mut fonts = FontsImpl::new(TextOptions::default(), FontDefinitions::default());
-
-            for halign in [Align::Min, Align::Center, Align::Max] {
-                for justify in [false, true] {
-                    for mut job in jobs() {
-                        job.halign = halign;
-                        job.justify = justify;
-
-                        let whole = GalleyCache::default().layout(
-                            &mut fonts,
-                            pixels_per_point,
-                            job.clone(),
-                            false,
-                        );
-
-                        let split = GalleyCache::default().layout(
-                            &mut fonts,
-                            pixels_per_point,
-                            job.clone(),
-                            true,
-                        );
-
-                        for (i, row) in whole.rows.iter().enumerate() {
-                            println!(
-                                "Whole row {i}: section_index_at_start={}, first glyph section_index: {:?}",
-                                row.row.section_index_at_start,
-                                row.row.glyphs.first().map(|g| g.section_index)
-                            );
-                        }
-                        for (i, row) in split.rows.iter().enumerate() {
-                            println!(
-                                "Split row {i}: section_index_at_start={}, first glyph section_index: {:?}",
-                                row.row.section_index_at_start,
-                                row.row.glyphs.first().map(|g| g.section_index)
-                            );
-                        }
-
-                        // Don't compare for equaliity; but format with a specific precision and make sure we hit that.
-                        // NOTE: we use a rather low precision, because as long as we're within a pixel I think it's good enough.
-                        similar_asserts::assert_eq!(
-                            format!("{:#.1?}", split),
-                            format!("{:#.1?}", whole),
-                            "pixels_per_point: {pixels_per_point:.2}, input text: '{}'",
-                            job.text
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_intrinsic_size() {
-        let pixels_per_point = [1.0, 1.3, 2.0, 0.867];
-        let max_widths = [40.0, 80.0, 133.0, 200.0];
-        let rounded_output_to_gui = [false, true];
-
-        for pixels_per_point in pixels_per_point {
-            let mut fonts = FontsImpl::new(TextOptions::default(), FontDefinitions::default());
-
-            for &max_width in &max_widths {
-                for round_output_to_gui in rounded_output_to_gui {
-                    for mut job in jobs() {
-                        job.wrap = TextWrapping::wrap_at_width(max_width);
-
-                        job.round_output_to_gui = round_output_to_gui;
-
-                        let galley_wrapped =
-                            layout(&mut fonts, pixels_per_point, job.clone().into());
-
-                        job.wrap = TextWrapping::no_max_width();
-
-                        let text = job.text.clone();
-                        let galley_unwrapped = layout(&mut fonts, pixels_per_point, job.into());
-
-                        let intrinsic_size = galley_wrapped.intrinsic_size();
-                        let unwrapped_size = galley_unwrapped.size();
-
-                        let difference = (intrinsic_size - unwrapped_size).length().abs();
-                        similar_asserts::assert_eq!(
-                            format!("{intrinsic_size:.4?}"),
-                            format!("{unwrapped_size:.4?}"),
-                            "Wrapped intrinsic size should almost match unwrapped size. Intrinsic: {intrinsic_size:.8?} vs unwrapped: {unwrapped_size:.8?}
-                                Difference: {difference:.8?}
-                                wrapped rows: {}, unwrapped rows: {}
-                                pixels_per_point: {pixels_per_point}, text: {text:?}, max_width: {max_width}, round_output_to_gui: {round_output_to_gui}",
-                            galley_wrapped.rows.len(),
-                            galley_unwrapped.rows.len()
-                            );
-                        similar_asserts::assert_eq!(
-                            format!("{intrinsic_size:.4?}"),
-                            format!("{unwrapped_size:.4?}"),
-                            "Unwrapped galley intrinsic size should exactly match its size. \
-                                {:.8?} vs {:8?}",
-                            galley_unwrapped.intrinsic_size(),
-                            galley_unwrapped.size(),
-                        );
-                    }
-                }
-            }
-        }
-    }
 
     #[test]
     fn test_fallback_glyph_width() {
