+++ crates/epaint/src/text/mod.rs
@@ -2,17 +2,28 @@
 
 pub mod cursor;
 mod font;
+mod font_data;
+mod font_definitions;
+mod font_id;
+mod font_tweak;
 mod fonts;
+mod galley_cache;
+mod glyph_rasterizer;
 mod index;
+mod styled_metrics;
 mod text_layout;
 mod text_layout_types;
+mod unicode;
 
 pub use {
-    fonts::{
-        FontData, FontDefinitions, FontFamily, FontId, FontInsert, FontPriority, FontTweak,
-        FontVariationAxis, Fonts, FontsImpl, FontsView, GlyphRasterizer, GlyphRasterizerRequest,
-        GlyphSource, GlyphSourcePreference, HintingTarget, InsertFontFamily, MAX_GLYPH_SIZE,
-        RasterizedGlyph, SmoothHinting, default_glyph_source, has_emoji_presentation,
+    font_data::{FontData, FontVariationAxis},
+    font_definitions::{FontDefinitions, FontInsert, FontPriority, InsertFontFamily},
+    font_id::{FontFamily, FontId},
+    font_tweak::{FontTweak, HintingTarget, SmoothHinting},
+    fonts::{Fonts, FontsImpl, FontsView, MAX_GLYPH_SIZE},
+    glyph_rasterizer::{
+        GlyphRasterizer, GlyphRasterizerRequest, GlyphSource, GlyphSourcePreference,
+        RasterizedGlyph, default_glyph_source, has_emoji_presentation,
     },
     index::{ByteIndex, ByteRange, ByteRangeExt, CharIndex, CharRange, CharRangeExt},
     text_layout::*,
